backend/pyproject.toml
[tool.poetry]
name = "weatherapp"
version = "0.1.0"
description = "An app which displays weather information related to weather of a city"
authors = ["Shebeli"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^4.2.7"
requests = "^2.31.0"
djangorestframework = "^3.14.0"
django-debug-toolbar = "^4.2.0"
psycopg2-binary = "^2.9.9"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
